requirements.txt
# ==== Utilidades generales ====
requests
python-dotenv
pandas==2.2.2
numpy==2.1.3
tqdm==4.66.4
matplotlib==3.8.4
seaborn==0.13.2
psutil==5.9.8

# ==== NLP clásico / Word2Vec ====
spacy==3.8.7           # recuerda descargar el modelo con: python -m spacy download en_core_web_sm
gensim==4.3.3

# ==== Detección de idioma (FastText) ====
fasttext==0.9.2        # en Windows puede requerir toolchain; alternativa: fasttext-wheel==0.9.2

# ==== MathBERT / Transformers ====
torch==2.5.0           # CPU por defecto; si usas GPU instala la variante CUDA correspondiente
transformers==4.44.2
tokenizers==0.19.1
safetensors==0.4.3

# ==== ML utilitario ====
scikit-learn==1.5.2